requirements.txt.txt
torch==2.1.0
torchvision==0.16.0
numpy
scikit-learn
matplotlib
pillow
tqdm
pyyaml
